+++ app/src/main/java/com/example/grupo22_kotlin/presentation/navigation/AppNavigation.kt
@@ -1,27 +1,27 @@
 package com.example.grupo22_kotlin.presentation.navigation
 
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.example.grupo22_kotlin.presentation.screens.login.LoginScreen
-import com.example.grupo22_kotlin.presentation.screens.register.RegisterScreen
+
 import com.example.grupo22_kotlin.presentation.screens.signup.SignupScreen
+import com.example.grupo22_kotlin.presentation.screens.start.StartScreen
 
 @Composable
 fun AppNavigation(navController: NavHostController){
     NavHost(
         navController = navController,
-        startDestination = AppScreen.Login.route){
-        composable(route= AppScreen.Login.route) {
-            LoginScreen(navController)
+        startDestination = AppScreen.Start.route){
+        composable(route= AppScreen.Start.route) {
+            StartScreen(navController)
         }
         composable(route = AppScreen.Signup.route){
             SignupScreen()
         }
-        composable(route= AppScreen.Register.route){
-            RegisterScreen()
+        composable(route= AppScreen.Login.route){
+            LoginScreen()
         }
     }
