--- app/src/main/java/com/example/grupo22_kotlin/presentation/screens/register/RegisterScreen.kt
@@ -1,19 +1,0 @@
-package com.example.grupo22_kotlin.presentation.screens.register
-
-import android.annotation.SuppressLint
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-@Composable
-fun RegisterScreen() {
-    Scaffold(
-        topBar = {},
-        content = {
-            Text(text = "RegisterScreen")
-        },
-        bottomBar = {}
-
-    )
-}
