+++ app/src/main/java/com/example/grupo22_kotlin/presentation/screens/start/components/StartContent.kt
@@ -1,4 +1,4 @@
-package com.example.grupo22_kotlin.presentation.screens.login.components
+package com.example.grupo22_kotlin.presentation.screens.start.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
@@ -6,23 +6,20 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import com.example.grupo22_kotlin.R
 import com.example.grupo22_kotlin.presentation.components.DefaultButton
 import com.example.grupo22_kotlin.presentation.navigation.AppScreen
 
 @Composable
-fun LoginContent(navController: NavHostController){
+fun StartContent(navController: NavHostController){
     Column(
         modifier = Modifier.fillMaxWidth(),
         horizontalAlignment = Alignment.CenterHorizontally
@@ -40,7 +37,7 @@ fun LoginContent(navController: NavHostController){
         Spacer(modifier = Modifier.height(15.dp))
         Text(
             modifier = Modifier
-                .clickable { navController.navigate(route = AppScreen.Register.route) },
+                .clickable { navController.navigate(route = AppScreen.Login.route) },
             text = "I alredy have an account")
 
     }
